move Weights.h
