+++ Polygon_mesh_processing/include/CGAL/Polygon_mesh_processing/internal/fair_impl.h
@@ -24,7 +24,7 @@
 #include <map>
 #include <set>
 #include <CGAL/assertions.h>
-#include <CGAL/Polygon_mesh_processing/internal/Hole_filling/Weights.h>
+#include <CGAL/Polygon_mesh_processing/Weights.h>
 #include <CGAL/Timer.h>
 #include <CGAL/trace.h>
 #include <iterator>
